requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
PyMySQL==1.1.0
python-dotenv==1.0.0
bcrypt==4.0.1
Werkzeug==2.3.7